Change structure to facilitate finding center for each file
Previously, we found the center of the image only for the flats and used the center of the last flat for the center of each image. This code still does that, but is reorganized to allow COMP_IMAGE_GEOMETRY to find the center of the center wavelength of the background intensity.