Filter out 1083 with ND=8 from gifs
Add new routine COMP_GET_ND_FILTER to consistently get ND filter from a header correctly. It takes date_dir and wave_type besides the extension header because that will be a factor when we attempt to reprocess old data.

NDFILTER was not correctly getting passed to extensions in flat.fts.