Fix for versions of IDL before 8.2.2
kaula is running IDL 8.2. VALUE_LOCATE on versions of IDL before 8.2.2 can't handle input arrays of a single element. This means that when trying to find the nearest QU file, if there is only a single file to check to find the closest, the pipeline will crash. This should fix that.