Fix for misaligned flat fill
The non-annulus region of the flat can be filled via the flats/fill config parameter. If set, the fill mask was not done positioned correctly because the center/radius of the occulter/field was found for the distortion corrected flat, but a distortion uncorrected flat is actually used to do the flat correction. This commit saves another set of occulter/field center/radius and post angles in the flat FITS header that are used when computing the fill mask.