Fix for header passed to COMP_FIX_STRAY_LIGHT
The image header (with no geometry information in it) was passed to COMP_FIX_STRAY_LIGHT. This gave a nonsensical mask for COMP_FIX_STRAY_LIGHT to use. Now passing the corresponding flat header to COMP_FIX_STRAY_LIGHT which gives a much more reasonable fit.

But that is not really correct either, since we are finding centers for each image now, not just using the flat center. Unfortunately, we currently find the center in COMP_COMBINE_BEAMS, i.e., much later in the processing of each file.