Changing BACKGRND FITS keyword in primary extension to LCBKG
FITS readers complain about having FITS keywords with the same name in the primary and extension headers, so one of the names should be changed. Here, changing the primary keyword to LCBKG (for Line Center), but that is only slightly suggestive of the actual computation of the value. It is the median of the mean of the center 3 wavelengths in the annulus. So it is line-center-ish, but not exactly line center.